define all state machine actions on statusable model, get rid of additional entity and sinchronization problems, get ability for activerecord-like callbacks on changing status.
