+++ spec/controllers/options_controller_spec.rb
@@ -40,11 +40,11 @@ RSpec.describe OptionsController, type: :controller do
     context 'when not owner does it' do
       before(:each) { session[:user_id] = user.id + 1 }
 
-      it { is_expected.to be_success}
+      it { is_expected.to be_success }
 
       it 'redirect to root' do
         subject
-        expect(response.body).to eq("window.location = \"/\"")
+        expect(response.body).to eq('window.location = "/"')
       end
     end
   end
@@ -61,11 +61,11 @@ RSpec.describe OptionsController, type: :controller do
     context 'when not owner does it' do
       before(:each) { session[:user_id] = user.id + 1 }
 
-      it { is_expected.to be_success}
+      it { is_expected.to be_success }
 
       it 'redirect to root' do
         subject
-        expect(response.body).to eq("window.location = \"/\"")
+        expect(response.body).to eq('window.location = "/"')
       end
     end
   end
@@ -96,11 +96,11 @@ RSpec.describe OptionsController, type: :controller do
     context 'when not owner does it' do
       before(:each) { session[:user_id] = user.id + 1 }
 
-        it { is_expected.to be_success}
+      it { is_expected.to be_success }
 
       it 'redirect to root' do
         subject
-        expect(response.body).to eq("window.location = \"/\"")
+        expect(response.body).to eq('window.location = "/"')
       end
 
       it 'not change option title' do
